java: move java_dependency into a separate file and add Language::Java.java_home_env
Also update activemq to test against it
