+++ Library/Homebrew/language/java.rb
@@ -1,0 +1,8 @@
+module Language
+  module Java
+    def self.java_home_env(version=nil)
+      version_flag = " --version #{version}" if version
+      { :JAVA_HOME => "$(/usr/libexec/java_home#{version_flag})" }
+    end
+  end
+end
